+++ src/avp/types/physical_channel_id.rs
@@ -1,25 +1,25 @@
 use crate::avp::{QueryableAVP, WritableAVP};
 use crate::common::{Reader, ResultStr, Writer};
 
-const G_LENGTH: u16 = 4;
+const G_LENGTH: usize = 4;
 
 #[derive(Clone, Copy, Debug, PartialEq)]
 pub struct PhysicalChannelId {
-    pub data: [u8; G_LENGTH as usize],
+    pub data: [u8; G_LENGTH],
 }
 
 impl PhysicalChannelId {
-    const LENGTH: u16 = G_LENGTH;
+    const LENGTH: usize = G_LENGTH;
 
     pub fn try_read<'a, 'b>(reader: &'b mut impl Reader<'a>) -> ResultStr<Self> {
-        if reader.len() < Self::LENGTH as usize {
+        if reader.len() < Self::LENGTH {
             return Err("Incomplete PhysicalChannelId payload encountered");
         }
 
         Ok(Self {
             data: unsafe {
                 reader
-                    .peek_bytes(Self::LENGTH as usize)?
+                    .peek_bytes(Self::LENGTH)?
                     .try_into()
                     .unwrap_unchecked()
             },
@@ -28,7 +28,7 @@ impl PhysicalChannelId {
 }
 
 impl QueryableAVP for PhysicalChannelId {
-    fn get_length(&self) -> u16 {
+    fn get_length(&self) -> usize {
         Self::LENGTH
     }
 }
