+++ src/avp/types/proxy_authen_name.rs
@@ -19,8 +19,8 @@ impl ProxyAuthenName {
 }
 
 impl QueryableAVP for ProxyAuthenName {
-    fn get_length(&self) -> u16 {
-        self.data.len() as u16
+    fn get_length(&self) -> usize {
+        self.data.len()
     }
 }
