+++ src/avp/types/vendor_name.rs
@@ -19,8 +19,8 @@ impl VendorName {
 }
 
 impl QueryableAVP for VendorName {
-    fn get_length(&self) -> u16 {
-        self.data.len() as u16
+    fn get_length(&self) -> usize {
+        self.data.len()
     }
 }
